# specs/001-senam-backend-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SENAM API
  version: 0.1.0-mvp
  summary: SENAM backend API — MVP surface (Doha car-wash vertical, Arabic)
  description: |
    This document is the **planning sketch** of the SENAM public API. It enumerates
    the endpoint surface, the auth model, the error envelope, and the high-value
    request/response shapes. The **source of truth at runtime** is the OpenAPI
    document NestJS generates from `@nestjs/swagger` decorators (mounted at
    `/swagger` in non-production environments). A contract test in CI asserts that
    every path listed here exists in the generated document and vice versa
    (SC-006).

    Authentication: every protected endpoint requires a `Bearer` access token
    obtained via the email-OTP login flow. Refresh tokens are sent via the
    `/auth/refresh` endpoint.

    Errors: every non-2xx response follows RFC-7807 `application/problem+json`:
    `{ "type", "title", "status", "detail", "instance", "correlationId" }`.
servers:
  - url: https://api.senam.app/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /me/addresses/{id}:
    patch: { tags: [addresses], parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], responses: { "200": { description: ok } } }
    delete: { tags: [addresses], parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], responses: { "204": { description: deleted } } }
...
  /companies/{id}:
    get:
      tags: [companies]
      summary: Public company detail
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: ok } }
...
  /bookings/{id}/review:
    post:
      tags: [reviews]
      summary: Submit a review for a completed booking
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ratingCompany]
              properties:
                ratingCompany: { type: integer, minimum: 1, maximum: 5 }
                ratingStaff: { type: integer, minimum: 1, maximum: 5 }
                ratingSpeed: { type: integer, minimum: 1, maximum: 5 }
                ratingQuality: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
                photoObjectKeys: { type: array, items: { type: string } }
      responses: { "201": { description: Review created. } }
  /reviews/{id}/reply:
    post:
      tags: [reviews, provider]
      summary: One-time provider reply
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "201": { description: Reply created. }
        "409": { description: A reply already exists. }

  # ─── Notifications ────────────────────────────────────────────────────────
...
  /provider/applications:
    post:
      tags: [provider-applications]
      summary: Submit a new company application with KYC document references
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [legalName, displayName, slug, ownerEmail, ownerPassword, documents]
              properties:
                legalName: { type: string }
                displayName: { type: string }
                slug: { type: string }
                description: { type: string }
                phone: { type: string }
                ownerEmail: { type: string, format: email }
                ownerDisplayName: { type: string }
                ownerPassword: { type: string, minLength: 8 }
                documents:
                  type: array
                  minItems: 2
                  items:
                    type: object
                    required: [kind, objectKey]
                    properties:
                      kind:
                        type: string
                        enum: [commercial_registration, tax_card, owner_id, premises_photo, certification, insurance]
                      objectKey: { type: string }
      responses:
        "201": { description: Application submitted, pending admin review }
        "409": { description: slug or email already in use }
  /provider/applications/{id}:
    get:
      tags: [provider-applications]
      summary: Get the status of a submitted company application
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: ok }
        "404": { description: application_not_found }

  # ─── Provider dashboard ───────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...